Make OpenDirectory test helpers reusable
